+++ package.json
@@ -31,6 +31,7 @@
     "react-router": "^6.6.0",
     "react-router-dom": "^6.6.0",
     "react-spinners": "^0.13.7",
+    "react-spinners-kit": "^1.9.1",
     "react-spring": "^9.6.1",
     "react-three": "^0.9.7",
     "sass": "^1.57.1",
